backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.1
python-multipart
pytest
pytest-asyncio
httpx
aiosqlite
greenlet
